+++ src/lib/Form.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from "react";
+import { ReactNode, useEffect, useState } from "react";
 import { DeepPartial, FieldValues, Resolver, SubmitHandler, useForm, UseFormReturn } from "react-hook-form";
 import { jsonIsoDateReviver } from "./helpers/dateUtils";
 import { FormContext, FormContextProps } from "./context/FormContext";
@@ -46,6 +46,11 @@ interface FormProps<T extends FieldValues> {
    * the form ref
    */
   formRef?: React.MutableRefObject<HTMLFormElement | null>;
+
+  /**
+   * Prevents the form from being html-native submited by disabling the form till its client-side loaded
+   */
+  disableBeforeLoading?: boolean;
 }
 
 const Form = <T extends FieldValues>({
@@ -57,15 +62,24 @@ const Form = <T extends FieldValues>({
   disabled = false,
   autoSubmitConfig,
   formRef,
+  disableBeforeLoading = true,
 }: FormProps<T>) => {
   const revivedDefaultValues = defaultValues
     ? (JSON.parse(JSON.stringify(defaultValues), jsonIsoDateReviver) as DeepPartial<T>)
     : defaultValues;
 
+  // prevents from sending the form wihtout being proberly loaded by react hook form
+  const [disabledLoading, setDisabledLoading] = useState<boolean>(disableBeforeLoading);
+  useEffect(() => {
+    if (disabledLoading) {
+      setDisabledLoading(false);
+    }
+  }, []);
+
   const formMethods = useForm<T>({ resolver, defaultValues: revivedDefaultValues });
   const autoSubmitHandler = useAutoSubmit({ onSubmit, formMethods, autoSubmitConfig });
   return (
-    <FormContext.Provider value={{ requiredFields, disabled, ...formMethods }}>
+    <FormContext.Provider value={{ requiredFields, disabled: disabled || disabledLoading, ...formMethods }}>
       <form
         ref={(elem) => {
           if (formRef) {
@@ -73,6 +87,7 @@ const Form = <T extends FieldValues>({
           }
         }}
         onSubmit={autoSubmitHandler}
+        method="POST"
       >
         {children instanceof Function ? children({ ...formMethods, disabled, requiredFields }) : children}
       </form>