+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `disableBeforeLoading` property (true as default) to `Form` component to prevent the form from being html-native submited by disabling the form till its client-side loaded
+
 ## [2.3.0] - 2023-12-12
 
 ### Added